requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.18.0
biopython>=1.83
requests>=2.31.0
spacy>=3.7.0
# After install run: python -m spacy download en_core_web_sm
